% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_noise_factors.R
\name{add_noise_factors}
\alias{add_noise_factors}
\title{Add Noise to Nuisance Factors}
\usage{
add_noise_factors(n, mean = 0, sd = 1, transform_fn = function(x) x)
}
\arguments{
\item{n}{Number of samples to generate.}

\item{mean}{Mean of the normal distribution.}

\item{sd}{Standard deviation of the normal distribution.}

\item{transform_fn}{A function applied to the normal draw (default is identity).}
}
\value{
A numeric vector of noisy factor values.
}
\description{
Applies normally-distributed noise to one or more factors using
a specified transformation function.
}
