% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports_rsm.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{ccd}
\alias{SO}
\alias{rsm}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{rsm}{\code{\link[rsm]{ccd}}, \code{\link[rsm]{rsm}}, \code{\link[rsm:FO]{SO}}}
}}

